api/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.5.0
astroid==3.2.4
black==24.8.0
click==8.1.7
coverage==7.6.1
databases==0.9.0
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fire==0.6.0
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
lxml==5.3.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.1
opencv-python-headless==4.10.0.84
packaging==24.1
pathspec==0.12.1
pdf2docx==0.5.8
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.2.7
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==1.1.1
pytest==8.3.3
pytest-cov==5.0.0
python-docx==1.1.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
starlette==0.38.5
termcolor==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
